refactor: move auth components into features with stubs
